# server/docs/swagger.yaml
host: 127.0.0.1:1323
info:
  contact:
    name: API Support
  description: WISMED Wisite-API Server. Updated@ 2022-09-03T11:33:31+10:00
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: WISMED WISITE API
  version: "1.0"
paths:
...
  /api/file/fileitems:
    get:
      consumes:
      - application/json
      parameters:
      - description: file ID (md5)
        in: query
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK - get fileitems successfully
        "400":
          description: Fail - incorrect query param id
        "404":
          description: Fail - not found
        "500":
          description: Fail - internal error
      security:
      - ApiKeyAuth: []
      summary: get fileitems by given path or id.
      tags:
      - File
  /api/file/pathcontent:
    get:
      consumes:
      - application/json
      parameters:
      - description: year-month, e.g. 2022-05
        in: query
        name: ym
        required: true
        type: string
      - description: group path, e.g. group1/group2/group3
        in: query
        name: gpath
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK - get content successfully
        "500":
          description: Fail - internal error
      security:
      - ApiKeyAuth: []
      summary: get content under specific path.
      tags:
      - File
  /api/file/upload-bodydata:
    post:
      consumes:
      - application/octet-stream
      parameters:
      - description: filename for uploading data from body
        in: query
        name: fname
        required: true
        type: string
      - description: note for uploading file
        in: query
        name: note
        type: string
      - description: 1st category for uploading file
        in: query
        name: group0
        type: string
      - description: 2nd category for uploading file
        in: query
        name: group1
        type: string
      - description: 3rd category for uploading file
        in: query
        name: group2
        type: string
      - description: file data for uploading
        format: binary
        in: body
        name: data
        required: true
        schema:
          type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK - return storage path
        "400":
          description: Fail - file param is incorrect
        "500":
          description: Fail - internal error
      security:
      - ApiKeyAuth: []
      summary: upload file action via body content.
      tags:
      - File
  /api/file/upload-formfile:
    post:
      consumes:
      - multipart/form-data
      parameters:
      - description: note for uploading file
        in: formData
        name: note
        type: string
      - description: 1st category for uploading file
        in: formData
        name: group0
        type: string
      - description: 2nd category for uploading file
        in: formData
        name: group1
        type: string
      - description: 3rd category for uploading file
        in: formData
        name: group2
        type: string
      - description: file path for uploading
        in: formData
        name: file
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: OK - return storage path
        "400":
          description: Fail - file param is incorrect
        "500":
          description: Fail - internal error
      security:
      - ApiKeyAuth: []
      summary: upload file action via form file input.
      tags:
      - File
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: authorization
    type: apiKey
swagger: "2.0"